+++ macOS/LocalPackages/DataBrokerProtection/Sources/DataBrokerProtection/macOS/Debug/DataBrokerRunCustomJSONViewModel.swift
@@ -18,6 +18,7 @@
 
 import Foundation
 import BrowserServicesKit
+import DataBrokerProtectionShared
 import Common
 import ContentScopeScripts
 import Combine
@@ -488,13 +489,6 @@ final class DataBrokerRunCustomJSONViewModel: ObservableObject {
 
     func appVersion() -> String {
         AppVersion.shared.versionNumber
-    }
-}
-
-final class FakeSleepObserver: SleepObserver {
-
-    func totalSleepTime() -> TimeInterval {
-        return 0
     }
 }
