+++ macOS/LocalPackages/DataBrokerProtection/Sources/DataBrokerProtection/macOS/UINative/DBPUIViewModel.swift
@@ -22,6 +22,7 @@ import WebKit
 import BrowserServicesKit
 import Common
 import os.log
+import DataBrokerProtectionShared
 
 protocol DBPUIScanOps: AnyObject {
     func updateCacheWithCurrentScans() async
