+++ macOS/LocalPackages/DataBrokerProtection/Tests/DataBrokerProtectionTests/DataBrokerOperationActionTests.swift
@@ -49,7 +49,6 @@ final class DataBrokerOperationActionTests: XCTestCase {
             operationAwaitTime: 0,
             stageCalculator: stageCalulator,
             pixelHandler: pixelHandler,
-            sleepObserver: FakeSleepObserver(),
             shouldRunNextStep: { true }
         )
 
@@ -75,7 +74,6 @@ final class DataBrokerOperationActionTests: XCTestCase {
             operationAwaitTime: 0,
             stageCalculator: stageCalulator,
             pixelHandler: pixelHandler,
-            sleepObserver: FakeSleepObserver(),
             shouldRunNextStep: { true }
         )
 
@@ -108,7 +106,6 @@ final class DataBrokerOperationActionTests: XCTestCase {
             operationAwaitTime: 0,
             stageCalculator: stageCalulator,
             pixelHandler: pixelHandler,
-            sleepObserver: FakeSleepObserver(),
             shouldRunNextStep: { true }
         )
 
@@ -139,7 +136,6 @@ final class DataBrokerOperationActionTests: XCTestCase {
             operationAwaitTime: 0,
             stageCalculator: stageCalulator,
             pixelHandler: pixelHandler,
-            sleepObserver: FakeSleepObserver(),
             shouldRunNextStep: { true }
         )
         sut.webViewHandler = webViewHandler
@@ -163,7 +159,6 @@ final class DataBrokerOperationActionTests: XCTestCase {
             operationAwaitTime: 0,
             stageCalculator: stageCalulator,
             pixelHandler: pixelHandler,
-            sleepObserver: FakeSleepObserver(),
             shouldRunNextStep: { true }
         )
         emailService.shouldThrow = true
@@ -193,7 +188,6 @@ final class DataBrokerOperationActionTests: XCTestCase {
             clickAwaitTime: 0,
             stageCalculator: stageCalulator,
             pixelHandler: pixelHandler,
-            sleepObserver: FakeSleepObserver(),
             shouldRunNextStep: { true }
         )
         sut.webViewHandler = webViewHandler
@@ -214,7 +208,6 @@ final class DataBrokerOperationActionTests: XCTestCase {
             operationAwaitTime: 0,
             stageCalculator: stageCalulator,
             pixelHandler: pixelHandler,
-            sleepObserver: FakeSleepObserver(),
             shouldRunNextStep: { true }
         )
         sut.webViewHandler = webViewHandler
@@ -237,7 +230,6 @@ final class DataBrokerOperationActionTests: XCTestCase {
             operationAwaitTime: 0,
             stageCalculator: stageCalulator,
             pixelHandler: pixelHandler,
-            sleepObserver: FakeSleepObserver(),
             shouldRunNextStep: { true }
         )
         sut.webViewHandler = webViewHandler
@@ -261,7 +253,6 @@ final class DataBrokerOperationActionTests: XCTestCase {
             operationAwaitTime: 0,
             stageCalculator: stageCalulator,
             pixelHandler: pixelHandler,
-            sleepObserver: FakeSleepObserver(),
             shouldRunNextStep: { true }
         )
         let actionsHandler = ActionsHandler(step: step)
@@ -292,7 +283,6 @@ final class DataBrokerOperationActionTests: XCTestCase {
             operationAwaitTime: 0,
             stageCalculator: stageCalulator,
             pixelHandler: pixelHandler,
-            sleepObserver: FakeSleepObserver(),
             shouldRunNextStep: { true }
         )
         sut.webViewHandler = webViewHandler
@@ -316,7 +306,6 @@ final class DataBrokerOperationActionTests: XCTestCase {
             operationAwaitTime: 0,
             stageCalculator: stageCalulator,
             pixelHandler: pixelHandler,
-            sleepObserver: FakeSleepObserver(),
             shouldRunNextStep: { true }
         )
         sut.retriesCountOnError = 0
@@ -341,7 +330,6 @@ final class DataBrokerOperationActionTests: XCTestCase {
             operationAwaitTime: 0,
             stageCalculator: stageCalulator,
             pixelHandler: pixelHandler,
-            sleepObserver: FakeSleepObserver(),
             shouldRunNextStep: { true }
         )
         sut.webViewHandler = webViewHandler
@@ -361,7 +349,6 @@ final class DataBrokerOperationActionTests: XCTestCase {
             operationAwaitTime: 0,
             stageCalculator: stageCalulator,
             pixelHandler: pixelHandler,
-            sleepObserver: FakeSleepObserver(),
             shouldRunNextStep: { true }
         )
         sut.webViewHandler = webViewHandler
@@ -383,7 +370,6 @@ final class DataBrokerOperationActionTests: XCTestCase {
             operationAwaitTime: 0,
             stageCalculator: mockStageCalculator,
             pixelHandler: pixelHandler,
-            sleepObserver: FakeSleepObserver(),
             shouldRunNextStep: { true }
         )
 
@@ -404,7 +390,6 @@ final class DataBrokerOperationActionTests: XCTestCase {
             operationAwaitTime: 0,
             stageCalculator: mockStageCalculator,
             pixelHandler: pixelHandler,
-            sleepObserver: FakeSleepObserver(),
             shouldRunNextStep: { true }
         )
 
@@ -425,7 +410,6 @@ final class DataBrokerOperationActionTests: XCTestCase {
             operationAwaitTime: 0,
             stageCalculator: mockStageCalculator,
             pixelHandler: pixelHandler,
-            sleepObserver: FakeSleepObserver(),
             shouldRunNextStep: { true }
         )
 
@@ -446,7 +430,6 @@ final class DataBrokerOperationActionTests: XCTestCase {
             operationAwaitTime: 0,
             stageCalculator: mockStageCalculator,
             pixelHandler: pixelHandler,
-            sleepObserver: FakeSleepObserver(),
             shouldRunNextStep: { true }
         )
 
@@ -467,7 +450,6 @@ final class DataBrokerOperationActionTests: XCTestCase {
             operationAwaitTime: 0,
             stageCalculator: stageCalulator,
             pixelHandler: pixelHandler,
-            sleepObserver: FakeSleepObserver(),
             shouldRunNextStep: { true }
         )
 
@@ -490,7 +472,6 @@ final class DataBrokerOperationActionTests: XCTestCase {
             operationAwaitTime: 0,
             stageCalculator: stageCalulator,
             pixelHandler: pixelHandler,
-            sleepObserver: FakeSleepObserver(),
             shouldRunNextStep: { true }
         )
