+++ macOS/LocalPackages/DataBrokerProtectionShared/Sources/DataBrokerProtectionShared/Operations/DataBrokerProfileQueryOperationManager.swift
@@ -35,7 +35,6 @@ protocol OperationsManager {
                       notificationCenter: NotificationCenter,
                       runner: WebJobRunner,
                       pixelHandler: EventMapping<DataBrokerProtectionSharedPixels>,
-                      sleepObserver: SleepObserver?,
                       showWebView: Bool,
                       isImmediateOperation: Bool,
                       userNotificationService: DataBrokerProtectionUserNotificationService,
@@ -49,7 +48,6 @@ extension OperationsManager {
                       notificationCenter: NotificationCenter,
                       runner: WebJobRunner,
                       pixelHandler: EventMapping<DataBrokerProtectionSharedPixels>,
-                      sleepObserver: SleepObserver?,
                       userNotificationService: DataBrokerProtectionUserNotificationService,
                       isManual: Bool,
                       shouldRunNextStep: @escaping () -> Bool) async throws {
@@ -60,7 +58,6 @@ extension OperationsManager {
                                notificationCenter: notificationCenter,
                                runner: runner,
                                pixelHandler: pixelHandler,
-                               sleepObserver: sleepObserver,
                                showWebView: false,
                                isImmediateOperation: isManual,
                                userNotificationService: userNotificationService,
@@ -76,7 +73,6 @@ struct DataBrokerProfileQueryOperationManager: OperationsManager {
                                notificationCenter: NotificationCenter = NotificationCenter.default,
                                runner: WebJobRunner,
                                pixelHandler: EventMapping<DataBrokerProtectionSharedPixels>,
-                               sleepObserver: SleepObserver?,
                                showWebView: Bool = false,
                                isImmediateOperation: Bool = false,
                                userNotificationService: DataBrokerProtectionUserNotificationService,
@@ -88,7 +84,6 @@ struct DataBrokerProfileQueryOperationManager: OperationsManager {
                                        database: database,
                                        notificationCenter: notificationCenter,
                                        pixelHandler: pixelHandler,
-                                       sleepObserver: sleepObserver,
                                        showWebView: showWebView,
                                        isManual: isImmediateOperation,
                                        userNotificationService: userNotificationService,
@@ -100,7 +95,6 @@ struct DataBrokerProfileQueryOperationManager: OperationsManager {
                                          database: database,
                                          notificationCenter: notificationCenter,
                                          pixelHandler: pixelHandler,
-                                         sleepObserver: sleepObserver,
                                          showWebView: showWebView,
                                          userNotificationService: userNotificationService,
                                          shouldRunNextStep: shouldRunNextStep)
@@ -113,7 +107,6 @@ struct DataBrokerProfileQueryOperationManager: OperationsManager {
                                    database: DataBrokerProtectionRepository,
                                    notificationCenter: NotificationCenter,
                                    pixelHandler: EventMapping<DataBrokerProtectionSharedPixels>,
-                                   sleepObserver: SleepObserver?,
                                    showWebView: Bool = false,
                                    isManual: Bool = false,
                                    userNotificationService: DataBrokerProtectionUserNotificationService,
@@ -141,7 +134,7 @@ struct DataBrokerProfileQueryOperationManager: OperationsManager {
             let event = HistoryEvent(brokerId: brokerId, profileQueryId: profileQueryId, type: .scanStarted)
             try database.add(event)
 
-            let extractedProfiles = try await runner.scan(brokerProfileQueryData, stageCalculator: stageCalculator, pixelHandler: pixelHandler, sleepObserver: sleepObserver, showWebView: showWebView, shouldRunNextStep: shouldRunNextStep)
+            let extractedProfiles = try await runner.scan(brokerProfileQueryData, stageCalculator: stageCalculator, pixelHandler: pixelHandler, showWebView: showWebView, shouldRunNextStep: shouldRunNextStep)
             Logger.dataBrokerProtection.log("Extracted profiles: \(extractedProfiles)")
 
             if !extractedProfiles.isEmpty {
@@ -289,7 +282,6 @@ struct DataBrokerProfileQueryOperationManager: OperationsManager {
                                      database: DataBrokerProtectionRepository,
                                      notificationCenter: NotificationCenter,
                                      pixelHandler: EventMapping<DataBrokerProtectionSharedPixels>,
-                                     sleepObserver: SleepObserver?,
                                      showWebView: Bool = false,
                                      userNotificationService: DataBrokerProtectionUserNotificationService,
                                      shouldRunNextStep: @escaping () -> Bool) async throws {
@@ -349,7 +341,6 @@ struct DataBrokerProfileQueryOperationManager: OperationsManager {
                                     extractedProfile: extractedProfile,
                                     stageCalculator: stageDurationCalculator,
                                     pixelHandler: pixelHandler,
-                                    sleepObserver: sleepObserver,
                                     showWebView: showWebView,
                                     shouldRunNextStep: shouldRunNextStep)
